Add files via upload
The Setheum Margin Trading White Paper - In this paper, I will elaborate the economic model of the Setheum Margin Trading Protocol - SettIndex on the Setheum Network. After going through this paper, the reader will understand the economic model and technicalities behind the SettIndex Margin Trading Protocol, from the Liquidity Pool to the Margin Providers and the Hedge Net Positions.