Uploaded SettIndex Money Market Paper
In this paper, I will elaborate the economic model of the Setheum Integrated Money Market Protocol - SettInDEX (Setheum Built-In DEX) on the Setheum Network. After going through this white paper, one should understand the economic model and technicalities behind the SettIndex Money Market, from the what happens pre-trade to Liquidity Pools and Trading Returns. This will be an efficient paper to the point.